Enable other repos to be defined in aocp.dependencies
AoCP does not host all repos that are required, so this allows us to define
repos that are on other remotes. By default, if no 'remote' variable is present
in aocp.dependencies, it will be added as 'name='TheCollective/xxx'' and
'remote='github'' in aocp_manifest.xml. Else, if a 'remote' variable is defined
,it will use the defined remote instead (this is only possible if the remote
name is available in default.xml, if not defined, just set remote as 'github'
but remember to specify the full name, like 'TheMuppets/proprietary_vendor_
samsung') and it will set 'name=' without 'TheCollective/'.

Sample aocp.dependencies can be found here: http://paste.ubuntu.com/5991665/